remove intermediate RedirectionActivity
