+++ sdk-core/src/main/java/co/reachfive/identity/sdk/core/SessionUtils.kt
@@ -1,17 +1,18 @@
 package co.reachfive.identity.sdk.core
 
 import android.app.Activity
+import android.content.Intent
+import android.net.Uri
+import androidx.browser.customtabs.CustomTabsIntent
 import co.reachfive.identity.sdk.core.api.ReachFiveApi
 import co.reachfive.identity.sdk.core.api.ReachFiveApiCallback
 import co.reachfive.identity.sdk.core.models.AuthToken
 import co.reachfive.identity.sdk.core.models.ReachFiveError
 import co.reachfive.identity.sdk.core.models.SdkConfig
 import co.reachfive.identity.sdk.core.models.SdkInfos
-import co.reachfive.identity.sdk.core.models.requests.LoginProviderRequest
+import co.reachfive.identity.sdk.core.models.requests.AuthCodeRequest
 import co.reachfive.identity.sdk.core.models.requests.RefreshRequest
-import co.reachfive.identity.sdk.core.utils.Failure
-import co.reachfive.identity.sdk.core.utils.Success
-import co.reachfive.identity.sdk.core.utils.SuccessWithNoContent
+import co.reachfive.identity.sdk.core.utils.*
 
 internal interface SessionUtils {
     var defaultScope: Set<String>
@@ -29,57 +30,138 @@ internal interface SessionUtils {
         origin: String? = null,
         activity: Activity,
     )
-
-    fun logout(
-        successWithNoContent: SuccessWithNoContent<Unit>,
-        failure: Failure<ReachFiveError>
-    )
 }
 
 class SessionUtilsClient(
     private val reachFiveApi: ReachFiveApi,
     private val sdkConfig: SdkConfig,
-    private val webLauncher: RedirectionActivityLauncher,
-    private val socialLoginAuth: SocialLoginAuthClient,
 ) : SessionUtils {
     companion object {
+        const val REDIRECTION_REQUEST_CODE = 52558
         const val codeResponseType = "code"
+        private const val CODE_VERIFIER_KEY = "code_verifier"
+
+        fun isRedirectionRequestCode(code: Int): Boolean =
+            code == REDIRECTION_REQUEST_CODE
     }
 
     override var defaultScope: Set<String> = emptySet()
 
-    fun loginWithProvider(
-        provider: String,
-        authCode: String? = null,
-        providerAccessToken: String? = null,
-        origin: String? = null,
-        scope: Collection<String>,
+    internal fun handleAuthorizationCompletion(
+        intent: Intent,
         success: Success<AuthToken>,
         failure: Failure<ReachFiveError>
     ) {
-        val loginProviderRequest = LoginProviderRequest(
-            provider = provider,
-            clientId = sdkConfig.clientId,
-            code = authCode,
-            providerToken = providerAccessToken,
-            origin = origin,
-            scope = scope.joinToString(" ")
-        )
-        reachFiveApi
-            .loginWithProvider(loginProviderRequest, SdkInfos.getQueries())
-            .enqueue(
-                ReachFiveApiCallback(
-                    success = { it.toAuthToken().fold(success, failure) },
-                    failure = failure
+        val uri = intent.data ?: Uri.EMPTY
+        val error = uri.let { ReachFiveError.fromRedirectionResult(it) }
+
+        val authCode = uri.getQueryParameter("code")
+        val codeVerifier = intent.getStringExtra(CODE_VERIFIER_KEY)
+
+        if (error != null)
+            failure(error)
+        else {
+            if (authCode == null) failure(ReachFiveError.Unexpected)
+            else if (codeVerifier == null) failure(ReachFiveError.Unexpected)
+            else {
+                val authCodeRequest = AuthCodeRequest(
+                    clientId = sdkConfig.clientId,
+                    code = authCode,
+                    redirectUri = sdkConfig.scheme,
+                    codeVerifier = codeVerifier
                 )
-            )
+
+                reachFiveApi
+                    .authenticateWithCode(authCodeRequest, SdkInfos.getQueries())
+                    .enqueue(
+                        ReachFiveApiCallback(
+                            success = { it.toAuthToken().fold(success, failure) },
+                            failure = failure
+                        )
+                    )
+            }
+        }
     }
 
-    override fun logout(
+    override fun loginWithWeb(
+        scope: Collection<String>,
+        state: String?,
+        nonce: String?,
+        origin: String?,
+        activity: Activity,
+    ) {
+        val intent = prepareCustomTabIntent(scope, origin = origin, state = state, nonce = nonce)
+        activity.startActivityForResult(intent, REDIRECTION_REQUEST_CODE)
+    }
+
+    internal fun webSocialLogin(
+        activity: Activity,
+        provider: Provider,
+        scope: Collection<String>,
+        origin: String?,
+    ) {
+        val intent = prepareCustomTabIntent(scope, origin = origin, provider = provider.name)
+        activity.startActivityForResult(intent, REDIRECTION_REQUEST_CODE)
+    }
+
+    fun loginCallback(
+        activity: Activity,
+        scope: Collection<String>,
+        tkn: String,
+    ) {
+        val intent = prepareCustomTabIntent(scope, tkn)
+        activity.startActivityForResult(intent, REDIRECTION_REQUEST_CODE)
+    }
+
+    private fun prepareCustomTabIntent(
+        scope: Collection<String>,
+        tkn: String? = null,
+        provider: String? = null,
+        origin: String? = null,
+        state: String? = null,
+        nonce: String? = null,
+    ): Intent {
+        val redirectUri = sdkConfig.scheme
+        val pkce = PkceAuthCodeFlow.generate(redirectUri)
+
+        val maybeTkn = if (tkn != null) {
+            mapOf("tkn" to tkn)
+        } else emptyMap()
+        val maybeProvider = if (provider != null) {
+            mapOf("provider" to provider)
+        } else emptyMap()
+        val maybeOrigin = if (origin != null) {
+            mapOf("origin" to origin)
+        } else emptyMap()
+        val maybeNonce = if (nonce != null) {
+            mapOf("nonce" to nonce)
+        } else emptyMap()
+        val maybeState = if (state != null) {
+            mapOf("state" to state)
+        } else emptyMap()
+
+        val request: Map<String, String> = mapOf(
+            "client_id" to sdkConfig.clientId,
+            "redirect_uri" to redirectUri,
+            "response_type" to "code",
+            "scope" to formatScope(scope),
+            "code_challenge" to pkce.codeChallenge,
+            "code_challenge_method" to pkce.codeChallengeMethod
+        ) + SdkInfos.getQueries() + maybeTkn + maybeProvider + maybeOrigin + maybeNonce + maybeState
+
+        val url = reachFiveApi.authorize(request).request().url.toString()
+
+        val customTabsIntent = CustomTabsIntent.Builder().build().intent
+        customTabsIntent.data = Uri.parse(url)
+        customTabsIntent.putExtra(CODE_VERIFIER_KEY, pkce.codeVerifier)
+
+        return customTabsIntent
+    }
+
+    internal fun logout(
         successWithNoContent: SuccessWithNoContent<Unit>,
         failure: Failure<ReachFiveError>
     ) {
-        socialLoginAuth.logoutFromAll()
         reachFiveApi
             .logout(SdkInfos.getQueries())
             .enqueue(
@@ -116,16 +198,6 @@ class SessionUtilsClient(
         scope: Collection<String>,
         activity: Activity
     ) {
-        webLauncher.loginCallback(activity, scope, tkn)
-    }
-
-    override fun loginWithWeb(
-        scope: Collection<String>,
-        state: String?,
-        nonce: String?,
-        origin: String?,
-        activity: Activity,
-    ) {
-        webLauncher.loginWithWeb(activity, scope, state, nonce, origin)
+        loginCallback(activity, scope, tkn)
     }
 }
