+++ sdk-core/src/main/java/co/reachfive/identity/sdk/core/SocialLoginClient.kt
@@ -8,6 +8,7 @@ import co.reachfive.identity.sdk.core.ReachFive.Companion.TAG
 import co.reachfive.identity.sdk.core.api.ReachFiveApi
 import co.reachfive.identity.sdk.core.api.ReachFiveApiCallback
 import co.reachfive.identity.sdk.core.models.*
+import co.reachfive.identity.sdk.core.models.requests.LoginProviderRequest
 import co.reachfive.identity.sdk.core.utils.Failure
 import co.reachfive.identity.sdk.core.utils.Success
 
@@ -40,9 +41,10 @@ internal interface SocialLoginAuth {
     )
 }
 
-internal class SocialLoginAuthClient(
+class SocialLoginAuthClient(
     private val reachFiveApi: ReachFiveApi,
     private val sdkConfig: SdkConfig,
+    private val sessionUtils: SessionUtilsClient,
     private val providersCreators: List<ProviderCreator>,
 ) : SocialLoginAuth {
 
@@ -59,6 +61,43 @@ internal class SocialLoginAuthClient(
     private var providers: List<Provider> = emptyList()
 
     override fun getProviders(): List<Provider> = providers
+
+    fun webProviderAuth(
+        activity: Activity,
+        provider: Provider,
+        scope: Collection<String>,
+        origin: String?,
+    ) {
+        sessionUtils.webSocialLogin(activity, provider, scope, origin)
+    }
+
+    fun completeNativeProviderAuth(
+        provider: String,
+        authCode: String? = null,
+        providerAccessToken: String? = null,
+        origin: String? = null,
+        scope: Collection<String>,
+        success: Success<AuthToken>,
+        failure: Failure<ReachFiveError>
+    ) {
+        val loginProviderRequest = LoginProviderRequest(
+            provider = provider,
+            clientId = sdkConfig.clientId,
+            code = authCode,
+            providerToken = providerAccessToken,
+            origin = origin,
+            scope = scope.joinToString(" ")
+        )
+
+        reachFiveApi
+            .loginWithProvider(loginProviderRequest, SdkInfos.getQueries())
+            .enqueue(
+                ReachFiveApiCallback(
+                    success = { it.toAuthToken().fold(success, failure) },
+                    failure = failure
+                )
+            )
+    }
 
     internal fun onStop() = providers.forEach { it.onStop() }
 
@@ -130,14 +169,12 @@ internal class SocialLoginAuthClient(
             when {
                 nativeProvider != null -> nativeProvider.create(
                     config,
-                    sdkConfig,
-                    reachFiveApi,
+                    this,
                     context
                 )
                 webViewProvider != null -> webViewProvider.create(
                     config,
-                    sdkConfig,
-                    reachFiveApi,
+                    this,
                     context
                 )
                 else -> {
