+++ sdk-facebook/src/main/java/co/reachfive/identity/sdk/facebook/FacebookProvider.kt
@@ -5,11 +5,10 @@ import android.content.Context
 import android.content.Intent
 import co.reachfive.identity.sdk.core.Provider
 import co.reachfive.identity.sdk.core.ProviderCreator
-import co.reachfive.identity.sdk.core.SessionUtilsClient
-import co.reachfive.identity.sdk.core.api.ReachFiveApi
-import co.reachfive.identity.sdk.core.api.ReachFiveApiCallback
-import co.reachfive.identity.sdk.core.models.*
-import co.reachfive.identity.sdk.core.models.requests.LoginProviderRequest
+import co.reachfive.identity.sdk.core.SocialLoginAuthClient
+import co.reachfive.identity.sdk.core.models.AuthToken
+import co.reachfive.identity.sdk.core.models.ProviderConfig
+import co.reachfive.identity.sdk.core.models.ReachFiveError
 import co.reachfive.identity.sdk.core.utils.Failure
 import co.reachfive.identity.sdk.core.utils.Success
 import co.reachfive.identity.sdk.facebook.FacebookProvider.Companion.REQUEST_CODE
@@ -27,16 +26,16 @@ class FacebookProvider : ProviderCreator {
 
     override fun create(
         providerConfig: ProviderConfig,
-        sessionUtils: SessionUtilsClient,
+        socialLoginClient: SocialLoginAuthClient,
         context: Context
     ): Provider {
-        return ConfiguredFacebookProvider(providerConfig, sessionUtils)
+        return ConfiguredFacebookProvider(providerConfig, socialLoginClient)
     }
 }
 
 internal class ConfiguredFacebookProvider(
     private val providerConfig: ProviderConfig,
-    private val sessionUtils: SessionUtilsClient,
+    private val socialLoginClient: SocialLoginAuthClient,
 ) : Provider {
     override val requestCode: Int = REQUEST_CODE
     override val name: String = FacebookProvider.NAME
@@ -67,7 +66,7 @@ internal class ConfiguredFacebookProvider(
             .registerCallback(callbackManager, object : FacebookCallback<LoginResult> {
                 override fun onSuccess(result: LoginResult) {
                     val accessToken = result.accessToken.token
-                    sessionUtils.loginWithProvider(
+                    socialLoginClient.completeNativeProviderAuth(
                         name,
                         providerAccessToken = accessToken,
                         origin = origin,
