+++ sdk-webview/src/main/java/co/reachfive/identity/sdk/webview/WebViewProvider.kt
@@ -4,10 +4,7 @@ import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import co.reachfive.identity.sdk.core.*
-import co.reachfive.identity.sdk.core.api.ReachFiveApi
-import co.reachfive.identity.sdk.core.api.ReachFiveApiCallback
 import co.reachfive.identity.sdk.core.models.*
-import co.reachfive.identity.sdk.core.models.requests.AuthCodeRequest
 import co.reachfive.identity.sdk.core.utils.Failure
 import co.reachfive.identity.sdk.core.utils.Success
 import co.reachfive.identity.sdk.webview.WebViewProvider.Companion.REQUEST_CODE
@@ -17,10 +14,10 @@ class WebViewProvider : ProviderCreator {
 
     override fun create(
         providerConfig: ProviderConfig,
-        sessionUtils: SessionUtilsClient,
+        socialLoginClient: SocialLoginAuthClient,
         context: Context
     ): Provider {
-        return ConfiguredWebViewProvider(providerConfig, )
+        return ConfiguredWebViewProvider(providerConfig, socialLoginClient)
     }
 
     companion object {
@@ -30,16 +27,13 @@ class WebViewProvider : ProviderCreator {
 
 internal class ConfiguredWebViewProvider(
     private val providerConfig: ProviderConfig,
-    private val redirectionHandler: RedirectionActivityResultHandler
+    private val socialLoginClient: SocialLoginAuthClient
 ) : Provider {
-
-    private val redirectionActivityLauncher = RedirectionActivityLauncher(sdkConfig, reachFiveApi)
-
     override val name: String = providerConfig.provider
     override val requestCode: Int = REQUEST_CODE
 
     override fun login(origin: String, scope: Collection<String>, activity: Activity) {
-        redirectionActivityLauncher.sloFlow(activity, this, scope, origin)
+        socialLoginClient.webProviderAuth(activity, this, scope, origin)
     }
 
     override fun onActivityResult(
@@ -48,7 +42,7 @@ internal class ConfiguredWebViewProvider(
         data: Intent?,
         success: Success<AuthToken>,
         failure: Failure<ReachFiveError>
-    ) = redirectionHandler.handle(resultCode, data, success, failure)
+    ) {}
 
     override fun toString(): String =
         providerConfig.provider
